chore(build): export the icons plugin direct from icons/plugin.js (#4640)
